fix: #4214 separate currency position pipe
