+++ apps/gauzy/src/app/@shared/pipes/index.ts
@@ -1,12 +1,13 @@
 import { DurationFormatPipe } from './duration-format.pipe';
 import { FilterArrayPipe } from './filter-array.pipe';
 import { Nl2BrPipe, TruncatePipe } from './text.pipe';
-import { ReplacePipe, CurrencyPositionPipe } from './replace.pipe';
+import { ReplacePipe } from './replace.pipe';
 import { TimeFormatPipe } from './time-format.pipe';
 import { UtcToLocalPipe } from './utc-to-local.pipe';
 import { DateTimeFormatPipe } from './datetime-format.pipe';
 import { DateFormatPipe } from './date-format.pipe';
 import { SafeHtmlPipe, SafeUrlPipe } from './safe/safe.pipe';
+import { CurrencyPositionPipe } from './currency-position.pipe';
 
 export * from './duration-format.pipe';
 export * from './filter-array.pipe';
@@ -17,6 +18,7 @@ export * from './datetime-format.pipe';
 export * from './date-format.pipe';
 export * from './utc-to-local.pipe';
 export * from './safe/safe.pipe';
+export * from './currency-position.pipe';
 
 export const Pipes = [
 	DurationFormatPipe,
@@ -29,6 +31,6 @@ export const Pipes = [
 	DateFormatPipe,
 	TruncatePipe,
 	UtcToLocalPipe,
-  SafeUrlPipe,
-  CurrencyPositionPipe
+  	SafeUrlPipe,
+  	CurrencyPositionPipe
 ];
